% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_edit}
\alias{demo_edit}
\title{Customer Credit Card Information}
\format{
\subsection{\code{demo_edit}}{

A data frame with 20 rows and 6 columns:
\describe{
\item{name}{Customer name}
\item{job}{Customer job}
\item{credit_card_provider}{Credit card provider}
\item{credit_card_security_code}{Credit card security code}
\item{date_obtained}{Date of obtaining the credit card}
\item{contactless_card}{Contactless card}
}
}
}
\source{
\url{https://CRAN.R-project.org/package=charlatan}
}
\usage{
demo_edit
}
\description{
A subset of fake customer credit card information inspired by the \code{{charlatan}} package.
}
\keyword{datasets}
