% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-data.R
\name{flter-data}
\alias{flter-data}
\alias{filter_data_ui}
\title{Shiny module to interactively filter a \code{data.frame}}
\usage{
filter_data_ui(id, show_nrow = TRUE, label_nrow = "Number of rows:")
}
\arguments{
\item{id}{Module id. See \code{\link[shiny]{callModule}}.}

\item{show_nrow}{Show number of filtered rows and total.}
}
\value{
A \code{list} with 2 elements :
\itemize{
\item \strong{data_filtered} : \code{\link[shiny]{reactive}} function returning data filtered.
\item \strong{code} : \code{\link[shiny]{reactiveValues}} with 2 slots :
\code{expr} (raw expression to filter data) and \code{dplyr} (code with dplyr pipeline).
}
}
\description{
Module generate inputs to filter \code{data.frame} according column's type.
Code to reproduce the filter is returned as an expression with filtered data.
}
