% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-globalenv.R
\name{get_data_packages}
\alias{get_data_packages}
\title{Get packages containing datasets}
\usage{
get_data_packages()
}
\value{
a character vector of packages names
}
\description{
Get packages containing datasets
}
\examples{
if (interactive()) {

  get_data_packages()

}
}
