% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-copypaste.R
\name{import-copypaste}
\alias{import-copypaste}
\alias{import_copypaste_ui}
\alias{import_copypaste_server}
\title{Import data with copy & paste}
\usage{
import_copypaste_ui(id)

import_copypaste_server(id, update_data = c("button", "always"))
}
\arguments{
\item{id}{Module's ID.}

\item{update_data}{When to update selected data:
\code{"button"} (when user click on button) or
\code{"always"} (each time user select a dataset in the list).}
}
\value{
\itemize{
\item UI: HTML tags that can be included in shiny's UI
\item Server: a \code{list} with one slot:
\itemize{
\item \strong{data}: a \code{reactive} function returning the selected \code{data.frame}.
}
}
}
\description{
Let the user copy data from Excel or text file then paste it into a text area to import it.
}
\examples{

library(shiny)
library(datamods)

ui <- fluidPage(
  tags$h3("Import data with copy & paste"),
  fluidRow(
    column(
      width = 4,
      import_copypaste_ui("myid")
    ),
    column(
      width = 8,
      tags$b("Imported data:"),
      verbatimTextOutput(outputId = "result")
    )
  )
)

server <- function(input, output, session) {

  imported <- import_copypaste_server("myid")

  output$result <- renderPrint({
    imported$data()
  })

}

if (interactive())
  shinyApp(ui, server)
}
