% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-file.R
\name{import-file}
\alias{import-file}
\alias{import_file_ui}
\alias{import_file_server}
\title{Import data from a file}
\usage{
import_file_ui(id)

import_file_server(id, update_data = c("button", "always"))
}
\arguments{
\item{id}{Module's id}

\item{update_data}{When to update selected data:
\code{"button"} (when user click on button) or
\code{"always"} (each time user select a dataset in the list).}
}
\value{
\itemize{
\item UI: HTML tags that can be included in shiny's UI
\item Server: a \code{list} with one slot:
\itemize{
\item \strong{data}: a \code{reactive} function returning the selected \code{data.frame}.
}
}
}
\description{
Let user upload a file and import data
}
\examples{


library(shiny)
library(datamods)

ui <- fluidPage(
  tags$h3("Import data from a file"),
  fluidRow(
    column(
      width = 4,
      import_file_ui("myid")
    ),
    column(
      width = 8,
      tags$b("Imported data:"),
      verbatimTextOutput(outputId = "result")
    )
  )
)

server <- function(input, output, session) {
  
  imported <- import_file_server("myid")
  
  output$result <- renderPrint({
    imported$data()
  })
  
}

if (interactive())
  shinyApp(ui, server)
}
