% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-globalenv.R
\name{import-globalenv}
\alias{import-globalenv}
\alias{import_globalenv_ui}
\alias{import_globalenv_server}
\title{Import data from Global Environment}
\usage{
import_globalenv_ui(id, default_choices = NULL)

import_globalenv_server(
  id,
  default_data = NULL,
  default_name = NULL,
  update_data = c("button", "always")
)
}
\arguments{
\item{id}{Module's ID.}

\item{default_choices}{Character vector of choices to use
if there's no \code{data.frame} in user's environment.}

\item{default_data}{Default \code{data.frame} to use.}

\item{default_name}{Default name to use.}

\item{update_data}{When to update selected data:
\code{"button"} (when user click on button) or
\code{"always"} (each time user select a dataset in the list).}
}
\value{

}
\description{
Let the user select a dataset from its own environment.
}
