% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-modal.R
\name{import-modal}
\alias{import-modal}
\alias{import_ui}
\alias{import_server}
\title{Get all import functions in a Modal}
\usage{
import_ui(id, from)

import_server(
  id,
  from = c("env", "file", "copypaste", "googlesheets", "database")
)
}
\arguments{
\item{id}{Module's id}

\item{from}{The import_ui & server to use, i.e. the method.
There are 5 options to choose from. ("env", "file", "copypaste", "googlsheets" & "database")}
}
\value{
\itemize{
\item UI: HTML tags that can be included in shiny's UI
\item Server: a \code{list} with one slot:
\itemize{
\item \strong{data}: a \code{reactive} function returning the selected \code{data.frame}.
}
}
}
\description{
Let the user choose a method of his choice to import data
}
\examples{

library(shiny)
library(datamods)

ui <- fluidPage(
  fluidRow(
    column(
      width = 12,
      tags$b("Imported data:"),
      verbatimTextOutput(outputId = "result")
    )
  )
)

server <- function(input, output, session) {
  
  imported <- import_server("myid", "env")
  output$result <- renderPrint({
    req(imported$data())
    imported$data()
  })
}

if (interactive())
  shinyApp(ui, server)
}
