% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-globalenv.R
\name{list_pkg_data}
\alias{list_pkg_data}
\title{List dataset contained in a package}
\usage{
list_pkg_data(pkg)
}
\arguments{
\item{pkg}{Name of the package, must be installed.}
}
\value{
a \code{character} vector or \code{NULL}.
}
\description{
List dataset contained in a package
}
\examples{

list_pkg_data("ggplot2")
}
