% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-group.R
\name{select-group}
\alias{select-group}
\alias{select_group_ui}
\alias{select_group_server}
\title{Select Group}
\usage{
select_group_ui(
  id,
  params,
  label = NULL,
  btn_label = "Reset filters",
  inline = TRUE
)

select_group_server(id, data_r, vars_r)
}
\arguments{
\item{id}{Module's id.}

\item{params}{A named list of parameters passed to each \code{selectizeInput}, you can use :
\code{inputId} (obligatory, must be variable name), \code{label}, \code{placeholder}.}

\item{label}{Character, global label on top of all labels.}

\item{btn_label}{Character, reset button label.}

\item{inline}{If \code{TRUE} (the default), select menus are horizontally positioned, otherwise vertically.}

\item{data_r}{Either a \code{\link[=data.frame]{data.frame()}} or a \code{\link[shiny:reactive]{shiny::reactive()}}
function returning a \code{data.frame} (do not use parentheses).}

\item{vars_r}{character, columns to use to create filters,
must correspond to variables listed in \code{params}. Can be a
\code{\link[shiny:reactive]{shiny::reactive()}} function, but values must be included in the initial ones (in \code{params}).}
}
\value{
a \code{\link[shiny:reactive]{shiny::reactive()}} function containing data filtered.
}
\description{
Group of mutually dependent select menus for filtering \code{data.frame}'s columns (like in Excel).
}
