% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-variables.R
\name{update-variables}
\alias{update-variables}
\alias{update_variables_ui}
\alias{update_variables_server}
\title{Select, rename and convert variables}
\usage{
update_variables_ui(id)

update_variables_server(id, data)
}
\arguments{
\item{id}{Module's ID}

\item{data}{a \code{data.frame}}
}
\value{

}
\description{
Select, rename and convert variables
}
