% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-opalr.R
\name{util_dataquieR_datatypes}
\alias{util_dataquieR_datatypes}
\title{convert opal data types to dataquieR compatible ones
see also: https://gitlab.com/umg_hgw/sc/implementation/opalrepository/-/issues/1}
\usage{
util_dataquieR_datatypes(opal_datatypes)
}
\arguments{
\item{opal_datatypes}{the opal data types vector}
}
\value{
the same vector with dataquieR compatible data types
}
\description{
convert opal data types to dataquieR compatible ones
see also: https://gitlab.com/umg_hgw/sc/implementation/opalrepository/-/issues/1
}
\author{
Jörg Henke
}
