% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-opalr.R
\name{util_ensure_default_label}
\alias{util_ensure_default_label}
\title{ensure that a column LABEL is there. OPAL provides label, but also localized ones such as label:de or label:en.}
\usage{
util_ensure_default_label(df_labelled, preferred_locale = "en")
}
\arguments{
\item{df_labelled}{the data frame with label columns}

\item{preferred_locale}{the prefered locale, en by default}
}
\value{
df_labelled enriched by a valid label column
}
\description{
ensure that a column LABEL is there. OPAL provides label, but also localized ones such as label:de or label:en.
}
\author{
Jörg Henke
}
