% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-opalr.R
\name{util_extract_missinglist}
\alias{util_extract_missinglist}
\title{extract the missings from the values}
\usage{
util_extract_missinglist(df_categories, ordered_variables)
}
\arguments{
\item{df_categories}{the OPAL categories sheet as a data frame}

\item{ordered_variables}{the variables in the correct order}
}
\value{
the full vector of missing lists ordered by the ordered_variables
}
\description{
extract the missings from the values
}
\author{
Jörg Henke
}
