% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-opalr.R
\name{util_prepare_categories}
\alias{util_prepare_categories}
\title{prepare the categories dataframe to have one line for all variables}
\usage{
util_prepare_categories(df_categories)
}
\arguments{
\item{df_categories}{the OPAL Categories data frame}
}
\value{
a data frame with OPAL Categories the dataquieR way
}
\description{
prepare the categories dataframe to have one line for all variables
}
\author{
Adrian Richter
}
