% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-column.R
\name{create-column}
\alias{create-column}
\alias{create_column_ui}
\alias{create_column_server}
\alias{list_allowed_operations}
\title{Create new column}
\usage{
create_column_ui(id)

create_column_server(
  id,
  data_r = reactive(NULL),
  allowed_operations = list_allowed_operations()
)

list_allowed_operations()
}
\arguments{
\item{id}{Module's ID.}

\item{data_r}{A \code{\link[shiny:reactive]{shiny::reactive()}} function returning a \code{data.frame}.}

\item{allowed_operations}{A \code{list} of allowed operations, see below for details.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} function returning the data.
}
\description{
This module allow to enter an expression to create a new column in \code{data.frame}.
}
\note{
User can only use a subset of function: (, +, -, *, ^, \%\%, \%/\%, /, ==, >, <, !=, <=, >=, &, |, abs, sign, sqrt, ceiling, floor, trunc, cummax, cummin, cumprod, cumsum, exp, expm1, log, log10, log2, log1p, cos, cosh, sin, sinh, tan, tanh, acos, acosh, asin, asinh, atan, atanh, cospi, sinpi, tanpi, gamma, lgamma, digamma, trigamma, round, signif, max, min, range, prod, sum, any, all, pmin, pmax, mean, paste, paste0, substr, nchar, trimws, gsub, sub, grepl, ifelse.
You can add more operations using the \code{allowed_operations} argument, for  example if you want to allow to use package lubridate, you can do:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{c(list_allowed_operations(), getNamespaceExports("lubridate"))
}\if{html}{\out{</div>}}
}
\examples{

library(shiny)
library(datamods)
library(reactable)

ui <- fluidPage(
  theme = bslib::bs_theme(version = 5L, preset = "bootstrap"),
  tags$h2("Create new column"),
  fluidRow(
    column(
      width = 4,
      create_column_ui("col")
    ),
    column(
      width = 8,
      reactableOutput(outputId = "table")
    )
  )
)

server <- function(input, output, session) {

  data_r <- create_column_server(
    id = "col",
    data_r = reactive(MASS::Cars93[, c(1, 3, 4, 5, 6, 10)])
  )

  output$table <- renderReactable({
    data <- req(data_r())
    reactable(
      data = data,
      bordered = TRUE,
      compact = TRUE,
      striped = TRUE
    )
  })
}

if (interactive())
  shinyApp(ui, server)
}
