% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-googlesheets.R
\name{import-googlesheets}
\alias{import-googlesheets}
\alias{import_googlesheets_ui}
\title{import data with a googlesheet link}
\usage{
import_googlesheets_ui(id)
}
\arguments{
\item{id}{Module's ID}
}
\value{
\itemize{
\item UI: HTML tags that can be included in shiny's UI
\item Server: a \code{list} with one slot:
\itemize{
\item \strong{data}: a \code{reactive} function returning the selected \code{data.frame}.
}
}
}
\description{
Let user paste link to a Google sheet then import the data.
}
