% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{module-echantillon}
\alias{module-echantillon}
\alias{sample_ui}
\alias{sample_server}
\title{Shiny Sample Module}
\usage{
sample_ui(id)

sample_server(id, my_data)
}
\arguments{
\item{id}{Module ID}

\item{my_data}{\code{reactive} containing a \code{data.frame} to use in the module.}
}
\description{
Shiny Sample Module
}
